7105 Move exception to correct package
